+++ client/src/scripts/others/game.ts
@@ -48,7 +48,6 @@ import { FindGameResult } from "common/scripts/config/config.ts";
 import { GameState, PlayArgs } from "./constants.ts";
 import { Scene2D } from "./scene.ts";
 import { JoinnedPacket } from "common/scripts/packets/joinned.ts";
-import { Wall } from "../objects/walls.ts";
 import { PingWorld } from "../objects/ping_world.ts";
 export class Game extends ClientGame<GameObject>{
     state:GameState=GameState.Idle