+++ client/src/scripts/uim/minimap.ts
@@ -50,6 +50,10 @@ export class MinimapModule extends UIModule<Game>{
     map_icons_name:Record<string,MinimapIconDef>={}
     map_icons:MinimapIconDef[]=[]
 
+    camera_x:number=0
+    camera_y:number=0
+
+    _mouseup!:(e:MouseEvent)=>void
     override on_init(): void {
         this.canvas=document.body.querySelector("#ui-map") as HTMLCanvasElement
         this.ctx=this.canvas.getContext("2d")!
@@ -59,6 +63,25 @@ export class MinimapModule extends UIModule<Game>{
         this.canvas.addEventListener("click",()=>{
             if(!this.fullscreen)this.toggle_fullscreen()
         })
+        this.canvas.addEventListener("mousedown",(e)=>{
+            if(e.button===2){
+                this.game.ui.begin_emote_wheel(this.game.input_manager.mouse_position,undefined,undefined,this.to_world(v2(e.clientX,e.clientY)),true)
+            }
+        })
+
+        this._mouseup=(e:MouseEvent)=>{
+            if(e.button===2){
+                this.game.ui.end_emote_wheel()
+            }
+        }
+        document.addEventListener("mouseup",this._mouseup)
+    }
+    to_world(position:Vec2):Vec2{
+        const rect=this.canvas.getBoundingClientRect()
+        const x=(position.x-rect.left-this.camera_x)/this.scale
+        const y=(position.y-rect.top-this.camera_y)/this.scale
+        const ms=this.game.minimap.meter_size
+        return v2(x/ms,y/ms)
     }
     async load_icons(){
         const base="/assets/img/menu/gui/map/"
@@ -71,6 +94,8 @@ export class MinimapModule extends UIModule<Game>{
         await this.load_map_icon("ping_airdrop","/assets/img/menu/gui/map/ping_airdrop.svg")
         await this.load_map_icon("ping_alert","/assets/img/menu/gui/map/ping_alert.svg")
         await this.load_map_icon("ping_here","/assets/img/menu/gui/map/ping_here.svg")
+        await this.load_map_icon("ping_heal","/assets/img/menu/gui/map/ping_heal.svg")
+        await this.load_map_icon("ping_gift","/assets/img/menu/gui/map/ping_gift.svg")
     }
 
     async load_map_icon(id:string,path:string){
@@ -141,6 +166,8 @@ export class MinimapModule extends UIModule<Game>{
         }
         this.scale=scale
 
+        this.camera_x=cameraX
+        this.camera_y=cameraY
         ctx.save()
         ctx.translate(cameraX,cameraY)
         ctx.scale(scale,scale)
@@ -416,7 +443,9 @@ export class MinimapModule extends UIModule<Game>{
             HideElement(this.canvas)
         }
     }
-    override on_destroy():void{}
+    override on_destroy():void{
+        document.removeEventListener("mouseup",this._mouseup)
+    }
     override on_clear():void{
         this.pings.length=0
         this.zones.length=0